+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 
+* Remove metadata from template as it is unsused by Tailor (#74).
 * Ignore selector (e.g. set in a `Tailorfile`) when a specific resource is
   given as Openshift does not accept both selector and resource (#72).
 * Surface errors during export as template (#70).